Update eos_download.py
Changed to allow downloading multiple files in one go which also allows to create an image bundle with an older EOS image along with a newer version of TerminAttr. Still need to add in some logic to prevent adding an EOS image which needs a newer TerminAttr swix in addition prior to importing into CVP. Also need to think how to cope with 2 image bundles which both use the same TerminAttr image as the imageUpload.py script will crap out if either image exists in CVP already.